MynoteBack/requirements.txt
# Mynote Backend 依赖列表
# Python 3.13+ 兼容

# ============================================
# 核心框架
# ============================================
fastapi==0.104.1              # Web 框架
uvicorn[standard]==0.24.0     # ASGI 服务器（必需）
pydantic==2.10.3              # 数据验证
pydantic-settings==2.6.1      # 配置管理

# ============================================
# 数据库
# ============================================
sqlalchemy==2.0.35            # ORM（兼容 Python 3.13）
alembic==1.12.1               # 数据库迁移（可选）

# ============================================
# HTTP 客户端
# ============================================
httpx==0.28.1                 # 异步 HTTP 客户端（用于 Suno API）
requests==2.32.3              # 同步 HTTP（保留兼容性）

# ============================================
# 工具库
# ============================================
python-dotenv==1.0.0          # 环境变量加载
python-multipart==0.0.6       # 文件上传支持（可选）
passlib[bcrypt]==1.7.4        # 密码加密（用于用户认证）
email-validator==2.3.0        # 邮箱验证

# ============================================
# 旧依赖（待废弃）
# ============================================
# Flask==3.0.0                # ❌ 已废弃，使用 FastAPI
# flask-cors==4.0.0           # ❌ 已废弃，FastAPI 内置 CORS
